+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [8.14.1](https://github.com/e-Spirit/fsxa-pattern-library/compare/v8.14.0...v8.14.1) (2022-12-15)
+
+
+### Bug Fixes
+
+* **nuxtbuild:** remove navigation import using absolute path ([#111](https://github.com/e-Spirit/fsxa-pattern-library/issues/111)) ([2f794ae](https://github.com/e-Spirit/fsxa-pattern-library/commit/2f794ae916bc1ec9082a2bcb1385bfa118fc39b4))
+
+# [8.14.0](https://github.com/e-Spirit/fsxa-pattern-library/compare/v8.13.2...v8.14.0) (2022-12-14)
+
+
+### Features
+
+* **content-projection:** add new content projection routing ([#106](https://github.com/e-Spirit/fsxa-pattern-library/issues/106)) ([09ce0db](https://github.com/e-Spirit/fsxa-pattern-library/commit/09ce0dbb900550776debc24f100cc9c7de8c1ea3))
+
 ## [8.13.2](https://github.com/e-Spirit/fsxa-pattern-library/compare/v8.13.1...v8.13.2) (2022-11-23)